+++ Src/CSharpier/CodeFormatter.cs
@@ -38,7 +38,13 @@ public static class CodeFormatter
                     options.NewLineBeforeMembersInObjectInitializers,
                 NewLineBeforeMembersInAnonymousTypes = options.NewLineBeforeMembersInAnonymousTypes,
                 NewLineBetweenQueryExpressionClauses = options.NewLineBetweenQueryExpressionClauses,
-                UsePrettierStyleTrailingCommas = options.UsePrettierStyleTrailingCommas,
+
+                QualificationForField = options.QualificationForField,
+                QualificationForProperty = options.QualificationForProperty,
+                QualificationForMethod = options.QualificationForMethod,
+                QualificationForEvent = options.QualificationForEvent,
+
+                UsePrettierStyleTrailingCommas = options.UsePrettierStyleTrailingCommas
             },
             cancellationToken
         );