+++ Src/CSharpier/SyntaxPrinter/SyntaxNodePrinters/IdentifierName.cs
@@ -4,6 +4,13 @@ internal static class IdentifierName
 {
     public static Doc Print(IdentifierNameSyntax node, FormattingContext context)
     {
-        return Token.Print(node.Identifier, context);
+        // TODO: Fabio: Check if parent is AssignmentExpressionSyntax SimpleAssignmentExpression Value = 2
+        // node.Declaration.Variables[0].ToString()
+
+        return Doc.Group(
+            "ciao.",
+            Token.Print(node.Identifier, context)
+        );
+        //return Token.Print(node.Identifier, context);
     }
 }